+++ projects/ngx-mat-lib/src/lib/xmat-models/xmat-array-item-from-object.model.ts
@@ -1,4 +1,0 @@
-export interface XmatArrayItemFromObject {
-    key: string;
-    value: any;
-}