+++ projects/ngx-mat-lib/src/lib/xmat-models/xmat-breakpoints.model.ts
@@ -1,0 +1,7 @@
+export enum XmatBreakpointKeys {
+    xs = "xs",
+    sm = "sm",
+    md = "md",
+    lg = "lg",
+    xl = "xl"
+}