+++ projects/ngx-mat-lib/src/lib/xmat-pipes/xmat-object-to-array.pipe.ts
@@ -1,24 +1,23 @@
-import {Pipe, PipeTransform} from "@angular/core";
+import { Pipe, PipeTransform, Injectable } from "@angular/core";
 //
-import {XmatArrayItemFromObject, XmatGenericObject} from "../xmat-models/index";
+import { XmatSelect, XmatGenericObject } from "../xmat-models/index";
 //
-import {forEach} from "lodash";
+import { forEach } from "lodash";
 
-
-@Pipe({name: "objectToArray"})
+@Injectable()
+@Pipe({ name: "objectToArray", pure: !1 })
 export class XmatObjectToArrayPipe implements PipeTransform {
-
-    transform(object: XmatGenericObject, limitTo: number = -1): XmatArrayItemFromObject[] {
-        if (!object || object !== Object(object) || Array.isArray(object)) {
-            return void 0;
-        }
-        !!limitTo || (limitTo = Object.keys(object).length);
-        const keys = [];
-        forEach(Object.keys(object), (key, index) => {
-            keys.push({key: key, value: object[key]});
-            // Breaks when limit is reached
-            return index > limitTo;
-        });
-        return keys;
+  transform<T>(object: XmatGenericObject<T>, limitTo?: number): XmatSelect<T>[] {
+    if (!object || object !== Object(object) || Array.isArray(object)) {
+      return void 0;
     }
+    (limitTo !== void 0) || (limitTo = Object.keys(object).length);
+    const keys = [];
+    forEach(Object.keys(object), (key, index) => {
+      keys.push({ key: key, value: object[key] });
+      // Breaks when limit is reached
+      return index < limitTo - 1;
+    });
+    return keys;
+  }
 }