+++ src/app/views/xmat-home/xmat-home.component.ts
@@ -36,6 +36,7 @@ export class XmatHomeComponent implements OnInit {
     isUbiTimeInputRequired = !0;
     myTestList: XmatLegendItem[];
     myTestListB: XmatLegendItem[];
+    optSelection: string;
     testFormGroup: FormGroup;
     testTimeModel: XmatTime = new XmatTime(12, 15);
     xmatTimeCtrl: FormControl;