+++ src/environments/environment.ts
@@ -1,5 +1,5 @@
 
 export const xmatEnvironment = {
-    appVersion: "7.2.15",
+    appVersion: "7.2.18",
     production: !1
 };